:wrench: Copy over Chapter Six into Dialogic
This is the first successful copy with the Markdown converter from
the Marteau utility suite. Some modifications such as splitting
narration blocks and condensing lines where emphasis was included
was necessary. Overall, the process was much faster using this
approach than the older approach of copying line by line.

To make a successful starter, it is recommended to save the JSON
from the converter into a separate file, then copy the events over
from that file into the timeline's "events" key.

Dialogic also has a bug where the timeline may reverse the order
of the nodes if a developer clicks on a node prematurely before the
timeline finishes loading. It is recommended to wait for the
timeline to finish before clicking on any node. This has been noted
and is actively being fixed on the Dialogic repository. See the issue
for details: https://github.com/coppolaemilio/dialogic/issues/709.

^HOH-5 work Development 20m

Signed-off-by: Marquis Kurt <software@marquiskurt.net>